# fix: replace for..of to prevent v8 deoptimization