--- a/lib/SourceMapDevToolPlugin.js
+++ b/lib/SourceMapDevToolPlugin.js
@@ -96,7 +96,8 @@ class SourceMapDevToolPlugin {
 									return module || source;
 								});
 
-								for(const module of modules) {
+								for(let idx = 0; idx < modules.length; idx++) {
+									const module = modules[idx];
 									if(!moduleToSourceNameMapping.get(module)) {
 										moduleToSourceNameMapping.set(module, ModuleFilenameHelpers.createFilename(module, moduleFilenameTemplate, requestShortener));
 									}
@@ -121,7 +122,8 @@ class SourceMapDevToolPlugin {
 				});
 
 				// find modules with conflicting source names
-				for(const module of allModules) {
+				for(let idx = 0; idx < allModules.length; idx++) {
+					const module = allModules[idx];
 					let sourceName = moduleToSourceNameMapping.get(module);
 					let hasName = conflictDetectionSet.has(sourceName);
 					if(!hasName) {