--- a/lib/dependencies/HarmonyModulesHelpers.js
+++ b/lib/dependencies/HarmonyModulesHelpers.js
@@ -34,8 +34,8 @@ class HarmonyModulesHelpers {
 		const desc = depInQuestion.describeHarmonyExport();
 		if(!desc.exportedName) return true;
 		let before = true;
-		for(const moduleDependency of module.dependencies) {
-			const dep = moduleDependency;
+		for(let idx = 0; idx < module.dependencies.length; idx++) {
+			const dep = module.dependencies[idx];
 			if(dep === depInQuestion) {
 				before = false;
 				continue;