@sokra 2017-08-30
@bmeurer Is there any plan that would make `for of` loops faster than `for(;;)` loops in newer version? What's the recommendation on long term?