@bmeurer 2017-08-30
Long-term `for..of` should be almost as fast as `for`. `for..of` on `Array`s comes at an additional cost, because of the iteration protocol, so at least in baseline (i.e. in the interpreter) it will always be slower than a plain old `for` loop. But this will probably not be noticeable.

I'd recommend to go for `for..of` unless you really find that the `for` loop improves your overall performance significantly.