@filipesilva 2017-08-30
The difference in this case was 1.1s on a rebuild of a medium sized project. It seemed that this only happened on projects with a lot of ES modules so perhaps there is an alternative approach to reduce the impact. The tendency towards ESM should make the problem bigger over time though.

I do want to highlight that the problem isn't the inherent performance but rather the de-optimization in V8: https://github.com/GoogleChrome/devtools-docs/issues/53#issuecomment-319107156

It may go away in the new v8 but until users upgrade node this deopt will remain.